PR-10: extract backends/x11.rs (+ placeholder backends/linux_console.rs)
Behaviour-preserving relocation:
- backends/x11.rs (new): x11rb::atom_manager! block (per D11),
  X11State + impl, run_x11, query_x11_active_window.
- backends/linux_console.rs: placeholder only. run_linux_console_backend_task
  stays in supervisor/mod.rs because it takes BackendContext, which would
  create a backends -> supervisor cycle if moved. PR-12 will absorb it
  into a FocusBackend impl together with the other backend_task adapters.

main.rs: 3526 -> 3300 LOC. supervisor/mod.rs's run_x11 import updated
to crate::backends::x11::run_x11.

Verification: cargo build clean; cargo test --bin kanata-switcher 261/0.
